Additional knives: silver and copper.
These textures are PLACEHOLDER textures only. I am not a particularly
good artist, especially when it comes to colour and shading.

Copper and silver knives are always registered (relying on Mystical
World's ingot ore dictionary registration is impossible as it is an
event that is fired with the lowest priority -- the same item
registration event), but their recipes are conditional on a source of
copper existing.