"Lore"-ify the Runestone advancement name.
Contextually here, if Druidism is based at all on Celtic druids, then
the runic script used to write the Irish language: Ogham. This is also
referred to as "The Language of Trees", because the letters often took
their names from different types of trees, and were also inscribed upon
trees.