Include copper and silver knives in the guide.
These are available by default because this depends on Mystical World
which provides copper and silver. They could be included at the end of
the book if there was ever a concern that copper and silver might not be
available, as invalid recipes won't be rendered.

Honestly, I'd rather some type of separate conditional page for knives
as that way they could be properly ordered.

Regardless, I don't see silver and copper being removed from Mystical
World any time soon, so I don't see the point in not including them.