Potentially cut down on the number of network packets being transmitted.
a) the Player Data Capability is no longer marked dirty automatically,
but only if an effect has actually ticked or ended (and thus the
capability has potentially changed). This may be an issue if there's a
permanent effect that thus results in it being updated every tick.

b) The Grove Capability is only ever sent to the target player, not that
it was previously an issue.

c) The Player Data Capability is only sent to people within 128 blocks
of the player, rather than to everyone in every dimension. This should
allow for synchronisation to be maintainted relatively well without
flooding everyone everywhere with packets.