Some mortar tweaks; store only the pestle.
Specifically, when the mortar is full and you cannot insert any more
items, but you have the pestle in the offhand, attempt to craft the
recipe using the pestle.

This makes "automation" a little bit kinder in the sense that you can
just spam right click with a stack full of wheat, for example, and get
flour without having to swap hands back and forth.

Hopefully it has no impact on anything else.

Additionally, when there is a stored pestle inside of a mortar, don't
allow anything else to be inserted.