One of those commits that got away from me.
Wildwood bows, updated the rest of the tile entities to change state and
thereby trigger a packet update instead of relying on manual packet --
in addition, hid all of them behind world.isRemote checks as I feel like
it could be leaking.

Removed WorldGenBarrow and WorldGenHut; these are no longer used but are
accessible via the git history.