Fix #126: upgrading to apothecary pouch no longer voids items.
This was "implemented" but when we moved from the Pyre crafting system
the "postCraft" of the recipe was never actually called by the tile
entity.

Either way, I've re-implemented it, and it will only apply to
single-stack applications (as postCraft is only ever called when there
is only one item and that item can only stack up to 1).