Nerf the heal of wildwood armor. (#144)
Theoretically this is healing 1/2 a heart randomly every 60 ticks four
times -- one for each piece of equipped armor. I've instead increased it
so that the bonus requires you are wearing every piece, and that it is
only triggered by the chestplate.

To compensate I've increased the chance of it triggering.