Fix #190: allow extinguishing with non-vanilla bucket and configurable fluids.
By default any liquid is accepted if it has a temperature of or less
than the default (300). By default this means that water is accepted and
lave is not.

Unfortunately I still had to hard-code in an exception for the empty
bucket as the FluidBucketWrapper doesn't swap the itemstacks out for
you. Blurgh!

Currently untested with other bucket containers and presumes that they
correctly change their state when emptied (and don't need to be
replaced).