Major work on overhauling Summon Creatures (#257)
As it stands, you can (and will be able to with CraftTweaker) add
ingredient-based recipes for summoning creatures. You can already sneak
while using Runic Shears to get Life Essence which can be used as a
summoning item (8 times).

The combinations of items for default recipes will probably be 1-3 items
(requiring 1-3 offering plates for the ritual). I've yet to decide
exactly which items will be used...

Ideas thus far: something involving seeds/eggs with chickens, something
involving wool for sheep, something involving wheat for cows, something
involving beetroot and/or apples for pigs, etc.

The obvious intent is to aim for materials (by default) that don't
require slaughtering animals. Likewise, as this is intended to be an
early ritual (and potentially a useful SkyBlock ritual), I'm making the
presumption that you don't yet have access to Animal Harvest.

Conflicts:
	src/main/java/epicsquid/roots/command/CommandRitual.java
	src/main/java/epicsquid/roots/ritual/RitualBase.java
	src/main/java/epicsquid/roots/ritual/RitualRegistry.java
	src/main/java/epicsquid/roots/ritual/RitualSummonCreatures.java
	src/main/java/epicsquid/roots/util/RitualUtil.java