Handle potential edge-case of #369: imbuing results in null module.
I'm not actually able to replicate what was happening in the bug report,
but I've identified the point that was problematic being that, somehow,
the module registry was returning null for iron boots.

I am able to imbue things properly with iron boots...

I've added a null check and ensure that the itemstack is always removed
before attempting to add the module (which will never be added if null),
meaning that the itemstack will simply be dropped on the ground next to
it and the animation will play.

I may add an error message.