Natures Scythe now harvests flowers too
Why did @Davoleo do this?
Because despite both Autumn's Fall and Nature's Scythe do pretty similar things neither of the two harvested flowers, meaning that if someone wanted to clear a field full of grass and flowers they would have to remove the flowers manually.
I've implemented it on Nature's Scythe instead of Autumn's Fall because I think it fits more the theme of the spell and furthermore Nature's Scythe had become a bit underpowered also because it's way slower than the other spell