Work on #934: casting tools don't interact with entities, blocks.
This may have unintended consequences on other things, but it basically
prioritizes the staff being able to cast a spell.

This means you can be closer to a block and shatter it without it
opening (#934), but also allows you to cast spells on entities from
close range without interacting (villagers, horses, etc).