Fix #1060: incorrect blaze powder output.
Technically, this is incorrect output from all recipes: I was doing the
correct thing by duplicating the chance outputs for each iteration, but
then I was also multiplying the output of those as well.

A recipe with a 0.5 chance to produce 1 of an item should become two 0.5
chances to produce 1 of an item.

Multiplying makes it become 0.5 chance to produce 2 of an item. Adding
and multiplying makes it two 0.5 chances to produce 2 of an item.