Tweaks for #1065: gravity, velocity, crit.
Living arrows have 30% less gravity than normal arrows, making them more
accurate.

Living arrows only need to be charged to 80% for it to count as a
"critical", instead of 100%. In reality this works out to be just when
the image changes to the third sprite, rather than a little while after
it.

Finally, they get 10% additional velocity.

To compensate, I've reduced the output from the recipe from 3 arrows to
2. The recipe may need to be made more expensive through others means as
well.