Implement visuals for grove stone power.
"Show hitbox mode" debug now displays not only the range of the grove
stone for generators, but also highlights in green correct generators,
and in red any generator that is, for some reason, incorrect.

e.g., if you place a flower down without its radially symmetrical
equivalent in the radius of the fairy grove stone, it will highlight
that flower in red, and then also place a red box where its pair should
go.