Fix #1158: improve herb alerts.
In "small window" (the default size of Minecraft in my development
environment), it's "forced" to only do 2 columns. Otherwise, it
dynamically calculates the available space based on the size of the
screen divided into thirds.

That said, the maximum is 3 columns.

In 3 column mode we adjust based on the off-hand by lifting it up a row.