Potential solution for #1199: player moving during light drifter.
The issue is that the forward impulse is never decayed or canceled when
you cast the spell, meaning that on the client-side it appears as though
your body is continuing to move.

In reality, when the spell ends, the re-sync from the server puts it
back where it's supposed to be.

The solution here is to just stop jumping and set xxa/zza to 0.

Future issue: elytra-flying.